requirements.txt
praw>=7.7.0
insightface>=0.7.3
onnxruntime>=1.17.0
fastapi>=0.104.0
uvicorn>=0.24.0
redis>=5.0.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
python-multipart>=0.0.6
